[ACS-7965] Refactoring
