+++ .secrets.baseline
@@ -440,37 +440,6 @@
         "line_number": 72
       }
     ],
-    "tests/environment/docker-compose-mtls.yml": [
-      {
-        "type": "Secret Keyword",
-        "filename": "tests/environment/docker-compose-mtls.yml",
-        "hashed_secret": "953bf453cebf3913edbe172ec292c4ff81d27415",
-        "is_verified": false,
-        "line_number": 34
-      },
-      {
-        "type": "Secret Keyword",
-        "filename": "tests/environment/docker-compose-mtls.yml",
-        "hashed_secret": "70c0b0b6d5ce34449b4bbf857972ffd286d082d9",
-        "is_verified": false,
-        "line_number": 43
-      },
-      {
-        "type": "Secret Keyword",
-        "filename": "tests/environment/docker-compose-mtls.yml",
-        "hashed_secret": "5baa61e4c9b93f3f0682250b6cf8331b7ee68fd8",
-        "is_verified": false,
-        "line_number": 98,
-        "is_secret": false
-      },
-      {
-        "type": "Secret Keyword",
-        "filename": "tests/environment/docker-compose-mtls.yml",
-        "hashed_secret": "c81d43b4c59486d3654825628a85bc0335704270",
-        "is_verified": false,
-        "line_number": 242
-      }
-    ],
     "tests/environment/docker-compose-pipeline-all-amps.yml": [
       {
         "type": "Secret Keyword",
@@ -560,6 +529,36 @@
         "hashed_secret": "d033e22ae348aeb5660fc2140aec35850c4da997",
         "is_verified": false,
         "line_number": 157
+      }
+    ],
+    "tests/environment/docker-compose-solr-mtls.yml": [
+      {
+        "type": "Secret Keyword",
+        "filename": "tests/environment/docker-compose-solr-mtls.yml",
+        "hashed_secret": "953bf453cebf3913edbe172ec292c4ff81d27415",
+        "is_verified": false,
+        "line_number": 34
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "tests/environment/docker-compose-solr-mtls.yml",
+        "hashed_secret": "70c0b0b6d5ce34449b4bbf857972ffd286d082d9",
+        "is_verified": false,
+        "line_number": 43
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "tests/environment/docker-compose-solr-mtls.yml",
+        "hashed_secret": "5baa61e4c9b93f3f0682250b6cf8331b7ee68fd8",
+        "is_verified": false,
+        "line_number": 98
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "tests/environment/docker-compose-solr-mtls.yml",
+        "hashed_secret": "c81d43b4c59486d3654825628a85bc0335704270",
+        "is_verified": false,
+        "line_number": 242
       }
     ],
     "tests/environment/docker-compose-sync-service.yml": [
@@ -783,15 +782,6 @@
         "is_secret": false
       }
     ],
-    "tests/tas-mtls/src/test/resources/default-elasticsearch.properties": [
-      {
-        "type": "Secret Keyword",
-        "filename": "tests/tas-mtls/src/test/resources/default-elasticsearch.properties",
-        "hashed_secret": "5baa61e4c9b93f3f0682250b6cf8331b7ee68fd8",
-        "is_verified": false,
-        "line_number": 7
-      }
-    ],
     "tests/tas-mtls/src/test/resources/default.properties": [
       {
         "type": "Secret Keyword",
@@ -875,5 +865,5 @@
       }
     ]
   },
-  "generated_at": "2026-04-10T12:33:44Z"
+  "generated_at": "2026-04-13T10:04:43Z"
 }
