+++ tests/tas-mtls/src/test/java/org/alfresco/rest/MtlsTestProperties.java
@@ -2,11 +2,8 @@ package org.alfresco.rest;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.context.annotation.PropertySources;
 
 @Configuration
-@PropertySources({@PropertySource({"classpath:default.properties"})})
 public class MtlsTestProperties
 {
     @Value("${testClient.keystore.location}")
