+++ tests/tas-mtls/src/test/resources/default-elasticsearch.properties
@@ -1,16 +1,3 @@
-# dataprep related
-alfresco.scheme=https
-alfresco.server=localhost
-alfresco.port=8443
-
-testClient.keystore.location=${CI_WORKSPACE}/keystores/testClient/testClient.keystore
-testClient.keystore.password=password
-testClient.keystore.type=JCEKS
-testClient.truststore.location=${CI_WORKSPACE}/keystores/testClient/testClient.truststore
-testClient.truststore.password=password
-testClient.truststore.type=JCEKS
-testClient.disableHostnameVerification=${HOSTNAME_VERIFICATION_DISABLED}
-
 # Elasticsearch mTLS endpoint for search engine health check
 searchEngine.mtls.url=https://localhost:9200
