+++ tests/tas-mtls/src/test/resources/default-solr.properties
@@ -1,0 +1,3 @@
+# Solr mTLS endpoint for search engine health check
+searchEngine.mtls.url=https://localhost:8083/solr
+
