+++ tests/tas-mtls/src/test/resources/default.properties
@@ -10,6 +10,3 @@ testClient.truststore.location=${CI_WORKSPACE}/keystores/testClient/testClient.t
 testClient.truststore.password=password
 testClient.truststore.type=JCEKS
 testClient.disableHostnameVerification=${HOSTNAME_VERIFICATION_DISABLED}
-
-# Solr mTLS endpoint for search engine health check
-searchEngine.mtls.url=https://localhost:8083/solr
