+++ tests/tas-mtls/src/test/resources/mtls-solr-suite.xml
@@ -1,17 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="Mutual TLS REST API" preserve-order="true">
+<suite name="Solr Mutual TLS REST API" preserve-order="true">
    <listeners>
       <listener class-name="org.alfresco.utility.report.HtmlReportListener"/>
       <listener class-name="org.alfresco.utility.testrail.TestRailExecutorListener"/>
       <listener class-name="org.alfresco.utility.testng.OSTestMethodSelector"/>
    </listeners>
 
-   <test name="Mutual TLS tests">
+   <test name="Solr Mutual TLS tests">
       <packages>
-         <package name="org.alfresco.rest.*">
-            <exclude name="org.alfresco.rest.mtls.elasticsearch.*"/>
-         </package>
+         <package name="org.alfresco.rest.mtls.solr.*"/>
       </packages>
    </test>
 </suite>
